Return transpose to numerical-utilities
When I moved this in the prior commit, I mentally asked a question to
myself: Why did Tamas' commit moving it here say "Many other types of
objects can be transposed, and it belongs in num-utils".  I couldn't
think of any, until I tried to recompile.  One example of something
that can be transposed is a data frame.